pkg/config: extract routing strict validators into compiler_validate_strict_routing.go (#4405)
Pure code-motion split of the 6,997-LOC compiler_validate_strict.go
(#4405). Moves the BGP/OSPF/router-id, FRR-auth, export, community, rib-group and route-filter strict commit-time validators
(validateXxxStrict) and their domain-local helpers verbatim into
compiler_validate_strict_routing.go (package config unchanged).

No validator logic, signature, or name changed: declarations are moved
byte-for-byte and only the per-file import set is recomputed. Verified by
the sorted top-level declaration-name set staying identical across the
split, and by go build ./... plus go test ./pkg/config/... staying green
at this step.
